requirements.txt
tensorflow
opencv-python
scikit-learn
matplotlib
pillow
requests
streamlit
python-dotenv
openai